refactor: move navigation hook to contexts directory
Consolidate navigation hook usage by moving it from components/Layout to contexts/NavigationContext for better organization and consistency across the application.
